posts_service/requirement.txt
SQLAlchemy
python-dotenv
psycopg2
grpcio
protobuf
grpcio-tools
